backend/requirements.txt
Flask==2.0.2
mysql-connector==2.2.9
boto3==1.20.26
Flask-Cors==3.0.10
